requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1

